src/server/requirements.txt
tensorflow
keras
numpy
pprint
flask
requests
python-dotenv
sklearn
flask
flask-cors
matplotlib
